modularizing the app
