+++ auth_service/Cargo.toml
@@ -12,3 +12,4 @@ bytes = "0.4.12"
 futures = "0.1.26"
 serde = { version = "1.0.90", features = ["derive"] }
 uuid = "0.7.4"
+sha2 = "0.8.0"
