+++ auth_service/src/user/user_view.rs
@@ -1,0 +1,8 @@
+
+
+struct UserView {
+    pub username: String,
+    pub email: String,
+    pub password: String
+}
+
